ai model/requirements.txt
tensorflow>=2.16
numpy>=1.26
pandas>=2.2
mysql-connector-python>=9.0
python-dotenv>=1.0
